requirements.txt
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
djoser
requests
cloudinary
python-decouple
psycopg2-binary
dj-database-url
drf-yasg
gunicorn
whitenoise
urllib3<2.0